Fixed compile

+++ src/transaction/transaction_group_complete_manager.cpp
@@ -55,7 +55,7 @@ namespace cubtx
     if (group_id < m_latest_closed_group_id)
       {
 	/* Already advanced to next group. No need to acquire mutex. */
-	er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete_mvcc: (%d)\n", group_id);
+	er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete_mvcc: (%lld)\n", group_id);
 	return;
       }
 
@@ -68,7 +68,7 @@ namespace cubtx
 	      {
 		/* Completed mvcc for group_id. No need to acquire mutex. */
 		assert (group_id <= m_latest_closed_group_id);
-		er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete_mvcc: (%d)\n", group_id);
+		er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete_mvcc: (%lld)\n", group_id);
 		return;
 	      }
 	  }
@@ -78,7 +78,7 @@ namespace cubtx
 	/* TODO - consider stop and optimize next call */
 	m_group_complete_condvar.wait (ulock, [&] {return is_group_mvcc_completed (group_id);});
 
-	er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete_mvcc: (%d)\n", group_id);
+	er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete_mvcc: (%lld)\n", group_id);
 	return;
       }
 #endif
@@ -88,7 +88,7 @@ namespace cubtx
     do_prepare_complete (thread_p);
     do_complete (thread_p);
     assert (is_group_mvcc_completed (group_id));
-    er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete_mvcc: (%d)\n", group_id);
+    er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete_mvcc: (%lld)\n", group_id);
   }
 
   //
@@ -99,7 +99,7 @@ namespace cubtx
     if (group_id < m_latest_closed_group_id)
       {
 	/* Already advanced to next group. No need to acquire mutex. */
-	er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete_logging: (%d)\n", group_id);
+	er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete_logging: (%lld)\n", group_id);
 	return;
       }
 
@@ -112,7 +112,7 @@ namespace cubtx
 	      {
 		/* Logging completed for group_id. No need to acquire mutex. */
 		assert (group_id <= m_latest_closed_group_id);
-		er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete_logging: (%d)\n", group_id);
+		er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete_logging: (%lld)\n", group_id);
 		return;
 	      }
 	  }
@@ -120,7 +120,7 @@ namespace cubtx
 	if (!need_wait_for_complete ())
 	  {
 	    /* Async case. */
-	    er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete_logging: (%d)\n", group_id);
+	    er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete_logging: (%lld)\n", group_id);
 	    return;
 	  }
 
@@ -129,7 +129,7 @@ namespace cubtx
 	/* TODO - consider stop and optimize next call */
 	m_group_complete_condvar.wait (ulock, [&] {return is_group_logged (group_id);});
 
-	er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete_logging: (%d)\n", group_id);
+	er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete_logging: (%lld)\n", group_id);
 	return;
       }
 #endif
@@ -139,7 +139,7 @@ namespace cubtx
     do_prepare_complete (thread_p);
     do_complete (thread_p);
     assert (is_group_logged (group_id));
-    er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete_logging: (%d)\n", group_id);
+    er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete_logging: (%lld)\n", group_id);
   }
 
   //
@@ -150,7 +150,7 @@ namespace cubtx
     if (group_id < m_latest_closed_group_id)
       {
 	/* Already advanced to next group. No need to acquire mutex. */
-	er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete: (%d)\n", group_id);
+	er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete: (%lld)\n", group_id);
 	return;
       }
 
@@ -163,7 +163,7 @@ namespace cubtx
 	      {
 		/* Group_id complete. No need to acquire mutex. */
 		assert (group_id <= m_latest_closed_group_id);
-		er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete: (%d)\n", group_id);
+		er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete: (%lld)\n", group_id);
 		return;
 	      }
 	  }
@@ -171,7 +171,7 @@ namespace cubtx
 	if (!need_wait_for_complete ())
 	  {
 	    /* Async case. */
-	    er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete: (%d)\n", group_id);
+	    er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete: (%lld)\n", group_id);
 	    return;
 	  }
 
@@ -179,7 +179,7 @@ namespace cubtx
 	std::unique_lock<std::mutex> ulock (m_group_complete_mutex);
 	/* TODO - consider stop and optimize next call */
 	m_group_complete_condvar.wait (ulock, [&] {return is_group_completed (group_id);});
-	er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete: (%d)\n", group_id);
+	er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete: (%lld)\n", group_id);
 
 	return;
       }
@@ -190,7 +190,7 @@ namespace cubtx
     do_prepare_complete (thread_p);
     do_complete (thread_p);
     assert (is_group_completed (group_id));
-    er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete: (%d)\n", group_id);
+    er_log_debug (ARG_FILE_LINE, "group_complete_manager::complete: (%lld)\n", group_id);
   }
 
   //
@@ -262,7 +262,7 @@ namespace cubtx
 	/* Close the current group. */
 	m_latest_closed_group_id.store (m_current_group_id);
 	m_latest_closed_group_state = GROUP_CLOSED;
-	er_log_debug (ARG_FILE_LINE, "group_complete_manager::close_current_group: (%d)\n", m_current_group_id);
+	er_log_debug (ARG_FILE_LINE, "group_complete_manager::close_current_group: (%lld)\n", (long long) m_current_group_id);
 
 	/* Advance with the group - copy and reinit it. */
 	m_current_group.transfer_to (m_latest_closed_group);
